CollisionManagerの削除

+++ 地球侵略/MoveGround.cpp
@@ -20,12 +20,12 @@ MoveGround::MoveGround(double x, double y, double moveUp, double moveDown, char 
 int MoveGround::update(const Collision & playerCol) {
 	//ここでは動くだけで当たり判定はせず，Player自身で判定と座標の処理を行う．
 	Move();
-	collision->updatePos(moveX, moveY);
+	collision->updatePos((int)moveX, (int)moveY);
 	return 0;
 }
 
 void MoveGround::Draw(int drawX, int drawY) {
-	DrawGraph(moveX - drawX, moveY - drawY, imgGround, true);
+	DrawGraph((int)moveX - drawX, (int)moveY - drawY, imgGround, true);
 
 	//当たり判定領域の表示
 	//DrawBox(collision->hitRange.xPos + collision->hitRange.xOffset - drawX, collision->hitRange.yPos + collision->hitRange.yOffset - drawY, collision->hitRange.xPos + collision->hitRange.xOffset + collision->hitRange.xSize - drawX, collision->hitRange.yPos + collision->hitRange.yOffset + collision->hitRange.ySize - drawY, 0xFF00FF, false);
